+++ cli.go
@@ -84,9 +84,8 @@ func DisplayUsage(out io.Writer, appName string, flagSet *flag.FlagSet, descript
 
 func doInit(out io.Writer, eout io.Writer, args []string) error {
 	var (
-		cName     string
-		dsnURI    string
-		storeType string
+		cName  string
+		dsnURI string
 	)
 	flagSet := flag.NewFlagSet("init", flag.ContinueOnError)
 	flagSet.BoolVar(&showHelp, "h", false, "help for init")
@@ -100,27 +99,12 @@ func doInit(out io.Writer, eout io.Writer, args []string) error {
 	case len(args) == 2:
 		cName, dsnURI = args[0], args[1]
 	case len(args) == 1:
-		cName = args[0]
+		cName, dsnURI = args[0], ""
 	default:
 		return fmt.Errorf("Expected: [OPTIONS] COLLECTION_NAME [DSN_URI], got %q", strings.Join(args, " "))
 	}
-	/*NOTE: excluding auto setting this inorder to prevent accidental
-	  environment contamination.
-		if dsnURI == "" {
-			dsnURI = os.Getenv("DATASET_DSN_URI")
-		}
-	*/
-	switch {
-	case strings.HasPrefix(dsnURI, "sqlite:"):
-		storeType = SQLSTORE
-	case strings.HasPrefix(dsnURI, "mysql:"):
-		storeType = SQLSTORE
-	case strings.HasPrefix(dsnURI, "pg:"):
-		storeType = SQLSTORE
-	default:
-		storeType = PTSTORE
-	}
-	c, err := Init(cName, dsnURI, storeType)
+	fmt.Printf("DEBUG cName %q, dsnURI %q\n", cName, dsnURI)
+	c, err := Init(cName, dsnURI)
 	if err == nil {
 		defer c.Close()
 	}